Kanteneinfärbung implementiert (und etwas mit NodeMap, was ich eigentlich schon mal gemacht hatte, was aber wohl irgendwie verloren geganzen ist)

+++ Knot3/Knot3-Implementierung/Knot3/KnotData/Direction.cs
@@ -115,12 +115,12 @@ namespace Knot3.KnotData
 			return v - d.Vector;
 		}
 
-		public static Vector3 operator / (Direction d, int i)
+		public static Vector3 operator / (Direction d, float i)
 		{
 			return d.Vector / i;
 		}
 
-		public static Vector3 operator * (Direction d, int i)
+		public static Vector3 operator * (Direction d, float i)
 		{
 			return d.Vector * i;
 		}
